requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver_manager==4.0.1
loguru==0.7.2
tqdm==4.66.1
requests==2.31.0
requests-toolbelt==1.0.0
lxml==5.4.0
undetected-chromedriver
Pillow
urllib3
python-docx